Bump version to 0.6.0
This version fixes a bug that caused cached endpoints to not have
the correct proxy_read_timeout.  The cached endpoints used the
default of 1m instead of whatever the user had configured in PE.

Now the cached endpoints set the proxy_read_timeout to the hiera
value that users can use to increase the timeout or defaults to
200.

Finally, the cached endpoints also set all the same headers as the
/ location does just so they are exactly the same as / but with
the added caching settings.